requirements.txt
SQLAlchemy==1.4.27
pytest==6.2.5
pika==1.1.0
flask=2.1.1
psycopg2=2.9.2
PyJWT==1.7.1
Werkzeug=2.1.1